Fix PTS calculation on rollover.
When a rollover was occuring the timeoffset value was incorrectly
calculated due to taking the minimum PTS from the new set of rollovers,
this lead to a very large delta on the PTS for the audio and led to
incorrect PTS values.

This happened to be fixed by the silent gap code just below, which set
the sample pts and dts to be the expected next value, but this now makes
it so on well-formed media, the else branch is essentially a no-op
because the delta is zero.